Add files via upload
The dataset I used in my project is downloaded from Kaggle website: https://www.kaggle.com/adammaus/predicting-churn-for-bank-customers. It contains 10,000 observations and 14 variables, including 13 independent variables and 1 binary dependent variable 'Exited' that records if a bank client churn or not churn. 
I firstly explored the dataset and visualized variables and correlations to check if the assumptions of logistic regression model are met. Then, I used variables to build a logistic regression model and the final accuracy is 81.5% and I also explore a random forest non-parameter model (accuracy is 86.85%) to compare with the logistic regression model, and further improvement will be made for both models after I learn more about predictive model building.